# openapi.yaml
openapi: 3.1.0
info:
  title: ThreatAssessor API
  description: 'Architecture threat modelling — MITRE ATT&CK mapping, MoE expert review,
    SOC detection (31 DETECT rules), AIVSS scoring, TA export bundle (ta-export/1.0),
    TA Brain KG, and TACO Agent (multi-hop routing chain).


    ## Authentication

    All endpoints (except /health and /dashboard) require TM-API-KEY: <key> request
    header.


    ## Typical workflows

    - Quick assessment: POST /api/v1/analyze → full threat model + MITRE TTPs (~30s)

    - Full pipeline: analyze → POST /api/v1/jobs/expert-review → poll GET /api/v1/jobs/{id}/status

    - CI/CD gate: POST /api/v1/governance/check (50ms, no LLM) → GET /api/v1/reports/{arch}/export
    → check gate.result

    - TACO query: POST /api/v1/taco/run-sync → HopChain (brain → workspace → harness
    routing)

    - Brain query: POST /api/v1/brain/query → pattern predictions with confidence

    - MCP: python -m mcp_server.server (stdio) or --transport streamable-http for
    remote agents'
  version: 2.7.0
  contact:
    name: ThreatAssessor
    url: http://localhost:8000/dashboard
  license:
    name: Proprietary
  x-mcp-server:
    stdio: python -m mcp_server.server
    streamable-http: python -m mcp_server.server --transport streamable-http
    tools: 17
paths:
...
  /api/v1/jobs/expert-review:
    post:
      tags:
      - jobs
      summary: Submit Expert Review
      description: 'Queue a FULL_MOE expert review for an existing architecture.


        Returns immediately with a job_id. Poll GET /api/v1/jobs/{job_id}/status

        to track progress and retrieve the result when status == ''completed''.'
      operationId: submit_expert_review_api_v1_jobs_expert_review_post
      parameters:
      - name: TM-API-KEY
        in: header
        required: false
        schema:
          type: string
          description: ThreatAssessor API key for authentication
          title: Tm-Api-Key
        description: ThreatAssessor API key for authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExpertReviewRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - ApiKeyAuth: []
  /api/v1/jobs/{job_id}/status:
    get:
      tags:
      - jobs
      summary: Get Job Status
      description: "Poll the status of a queued or running job.\n\nReturns:\n    status:\
        \   queued | running | completed | failed | blocked\n    progress: 0-100\n\
        \    message:  current stage description\n    result:   present when status\
        \ == 'completed'\n    error:    present when status == 'failed' or 'blocked'"
      operationId: get_job_status_api_v1_jobs__job_id__status_get
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          title: Job Id
      - name: TM-API-KEY
        in: header
        required: false
        schema:
          type: string
          description: ThreatAssessor API key for authentication
          title: Tm-Api-Key
        description: ThreatAssessor API key for authentication
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - ApiKeyAuth: []
  /api/v1/mcp/access-signals:
    get:
      tags:
      - jobs
      summary: Get Mcp Access Signals
      description: 'Return the current MCP session access pattern signals.


        Used by the dashboard MCP tab to power the DETECT-MCP-001/021/022 status cards

        and the tool-usage heatmap. No parameters — always returns the live state

        of the module-level MCPAccessLogger singleton.'
      operationId: get_mcp_access_signals_api_v1_mcp_access_signals_get
      parameters:
      - name: TM-API-KEY
        in: header
        required: false
        schema:
          type: string
          description: ThreatAssessor API key for authentication
          title: Tm-Api-Key
        description: ThreatAssessor API key for authentication
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - ApiKeyAuth: []
  /api/v1/mcp/jobs:
    get:
      tags:
      - jobs
      summary: List Mcp Jobs
      description: 'Return all non-expired jobs from the job store.


        Used by the dashboard MCP tab Jobs sub-pane. Jobs expire after 1 hour;

        this endpoint returns whatever is still in the in-memory store.'
      operationId: list_mcp_jobs_api_v1_mcp_jobs_get
      parameters:
      - name: TM-API-KEY
        in: header
        required: false
        schema:
          type: string
          description: ThreatAssessor API key for authentication
          title: Tm-Api-Key
        description: ThreatAssessor API key for authentication
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - ApiKeyAuth: []
...
components:
  schemas:
...
    ExpertReviewRequest:
      properties:
        arch_name:
          type: string
          title: Arch Name
        critic_mode:
          type: string
          title: Critic Mode
          default: partial_parallel
      type: object
      required:
      - arch_name
      title: ExpertReviewRequest
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: TM-API-KEY
servers:
- url: http://localhost:8000
  description: Local development
security:
- ApiKeyHeader: []
...